pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Blaze-Sudio"
version = "3.2.0"
description = "This is a really cool game/physics/graphics/application engine written in pygame!"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Tsunami014", email = "tsunami014@duck.com"}
]
maintainers = [
    {name = "Tsunami014", email = "tsunami014@duck.com"}
]
keywords = ["Python", "games", "engine"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows"
]
# Development Status options: 
# 1 - Planning
# 2 - Pre-Alpha
# 3 - Alpha
# 4 - Beta
# 5 - Production/Stable
# 6 - Mature
# 7 - Inactive

dependencies = []

requires-python = ">= 3.11"

[project.urls]
Homepage = "https://github.com/Tsunami014/Blaze-Sudio"
Download = "https://github.com/Tsunami014/Blaze-Sudio/archive/refs/tags/v3.2.0.tar.gz"
Documentation = "https://tsunami014.gitbook.io/blaze-sudios/"
Repository = "https://github.com/Tsunami014/Blaze-Sudio"
Issues = "https://github.com/Tsunami014/Blaze-Sudio/issues"

[project.scripts]
BlazeSudio = "BlazeSudio:main"

[project.optional-dependencies]
graphics = [
    "pysdl2",
    "pysdl2-dll",
    "numpy"
]

image = [
    "Pillow",
    "numpy",
    "scikit-image",
]

collisions = [
    "shapely",
]

game = [
    "Blaze-Sudio[graphics]",
    "Blaze-Sudio[collisions]",
    "scikit-image",
    "pyperclip",
]

all = [
    "Blaze-Sudio[image]",
    "Blaze-Sudio[game]",
]

[tool.setuptools.packages.find]
exclude = [
    "BlazeSudio.speed.setup",
]

[tool.setuptools.package-data]
"BlazeSudio" = [
    "elementGen/nodes/*.py",
    "ldtk/internal-icons.png",
    "graphics/defaultTheme/*",
    "graphics/GUI/textboxify/data/**",

    "**/*.pyi",
    "**/*.so", "**/*.pyd"
]
